+++ database/migrations/20250901142206_add_post_uri_to_messages_table.php
@@ -1,0 +1,21 @@
+<?php
+
+use Movim\Migration;
+use Illuminate\Database\Schema\Blueprint;
+
+class AddPostUriToMessagesTable extends Migration
+{
+    public function up()
+    {
+        $this->schema->table('messages', function (Blueprint $table) {
+            $table->longText('posturi')->nullable();
+        });
+    }
+
+    public function down()
+    {
+        $this->schema->table('messages', function (Blueprint $table) {
+            $table->dropColumn('posturi');
+        });
+    }
+}